Studnet can see their grades and final grade percentage. Closes #28

+++ LMS/UserInformation/UserModels/Submission.cs
@@ -11,6 +11,7 @@ namespace UserInformation.UserModels
         public int AssignmentId { get; set; }
         public string? Content { get; set; }
         public DateTime SubmissionDate { get; set; }
-        public int? Grade { get; set; }
+        public double? Grade { get; set; }
+
     }
 }
